tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-app-history-navigate-preventDefault.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  appHistory.oncurrentchange = t.unreached_func("currentchange should not fire");
  appHistory.onnavigate = e => e.preventDefault();
  await promise_rejects_dom(t, "AbortError", appHistory.navigate("#foo").committed);
}, "AppHistoryCurrentChangeEvent does not fire when onnavigate preventDefault() is called");
</script>
